More partials & Verb switch statement
- Moved check for updates code into a partial and own function
- Moved missing cores check code into a partial and own function
- All verbs are handled by a switch statement vs bools and a switch
